+++ src/plone/app/testing/bbb.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """Backwards-compatibility test class for PloneTestCase for Dexterity."""
 
 from AccessControl import getSecurityManager
@@ -15,7 +14,7 @@ def _createMemberarea(portal, user_id):
     mtool = portal.portal_membership
     members = mtool.getMembersFolder()
     if members is None:
-        _createObjectByType('Folder', portal, id='Members')
+        _createObjectByType("Folder", portal, id="Members")
     if not mtool.getMemberareaCreationFlag():
         mtool.setMemberareaCreationFlag()
     mtool.createMemberArea(user_id)
@@ -24,19 +23,19 @@ def _createMemberarea(portal, user_id):
 
 
 class PloneTestCaseFixture(testing.PloneSandboxLayer):
-
     defaultBases = (testing.PLONE_FIXTURE,)
 
     def setUpZope(self, app, configurationContext):
         import plone.app.contenttypes
+
         self.loadZCML(package=plone.app.contenttypes)
 
     def setUpPloneSite(self, portal):
         # restore default workflow
-        testing.applyProfile(portal, 'Products.CMFPlone:testfixture')
+        testing.applyProfile(portal, "Products.CMFPlone:testfixture")
 
         # add default content
-        testing.applyProfile(portal, 'plone.app.contenttypes:plone-content')
+        testing.applyProfile(portal, "plone.app.contenttypes:plone-content")
 
         # add home folder for default test user
         _createMemberarea(portal, testing.TEST_USER_ID)
@@ -44,20 +43,19 @@ class PloneTestCaseFixture(testing.PloneSandboxLayer):
 
 PTC_FIXTURE = PloneTestCaseFixture()
 PTC_FUNCTIONAL_TESTING = testing.FunctionalTesting(
-    bases=(PTC_FIXTURE,), name='PloneTestCase:Functional')
+    bases=(PTC_FIXTURE,), name="PloneTestCase:Functional"
+)
 
 
 class PloneTestCase(Functional, unittest.TestCase):
-
     layer = PTC_FUNCTIONAL_TESTING
 
     def setUp(self):
         """Set up before each test."""
         self.beforeSetUp()
-        self.app = self.layer['app']
-        self.portal = self.layer['portal']
-        self.folder = self.portal.portal_membership.getHomeFolder(
-            testing.TEST_USER_ID)
+        self.app = self.layer["app"]
+        self.portal = self.layer["portal"]
+        self.folder = self.portal.portal_membership.getHomeFolder(testing.TEST_USER_ID)
         transaction.commit()
         self.afterSetUp()
 
@@ -72,7 +70,7 @@ class PloneTestCase(Functional, unittest.TestCase):
         """Tear down after each test."""
         self.beforeTearDown()
         transaction.abort()
-        super(PloneTestCase, self).tearDown()
+        super().tearDown()
         self.afterTearDown()
 
     def beforeTearDown(self):
@@ -86,14 +84,14 @@ class PloneTestCase(Functional, unittest.TestCase):
         testing.setRoles(self.portal, name, roles)
 
     def setGroups(self, groups, name=testing.TEST_USER_ID):
-        '''Changes the user's groups.'''
-        uf = self.portal['acl_users']
+        """Changes the user's groups."""
+        uf = self.portal["acl_users"]
         uf.userSetGroups(name, list(groups))
         user = getSecurityManager().getUser()
         if name == user.getId():
             self.login(user.getUserName())
 
-    def setPermissions(self, permissions, role='Member'):
+    def setPermissions(self, permissions, role="Member"):
         """Changes the permissions assigned to role."""
         self.portal.manage_role(role, list(permissions))
 
@@ -103,7 +101,7 @@ class PloneTestCase(Functional, unittest.TestCase):
 
     def loginAsPortalOwner(self, userName=testing.SITE_OWNER_NAME):
         """Log in to the portal as the user who created it."""
-        zope.login(self.app['acl_users'], userName)
+        zope.login(self.app["acl_users"], userName)
 
     def logout(self):
         """Log out, i.e. become anonymous."""