+++ src/plone/app/testing/interfaces.py
@@ -1,17 +1,18 @@
-# -*- coding: utf-8 -*-
 # Constants for names used in layer setup - also importable from
 # plone.app.testing directly
 
-PLONE_SITE_ID = 'plone'
-PLONE_SITE_TITLE = u'Plone site'
-DEFAULT_LANGUAGE = 'en'
+PLONE_SITE_ID = "plone"
+PLONE_SITE_TITLE = "Plone site"
+DEFAULT_LANGUAGE = "en"
 
-TEST_USER_NAME = 'test-user'
-TEST_USER_ID = 'test_user_1_'
-TEST_USER_PASSWORD = 'correct horse battery staple'
-TEST_USER_ROLES = ['Member', ]
+TEST_USER_NAME = "test-user"
+TEST_USER_ID = "test_user_1_"
+TEST_USER_PASSWORD = "correct horse battery staple"
+TEST_USER_ROLES = [
+    "Member",
+]
 
-SITE_OWNER_NAME = 'admin'
-SITE_OWNER_PASSWORD = 'secret'
+SITE_OWNER_NAME = "admin"
+SITE_OWNER_PASSWORD = "secret"
 
 ROBOT_TEST_LEVEL = 5