+++ src/plone/app/testing/selenium_layers.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 from plone.app.testing import FunctionalTesting
 from plone.app.testing import PLONE_FIXTURE
 from plone.app.testing import TEST_USER_NAME
@@ -11,32 +10,33 @@ import transaction
 
 
 class SeleniumLayer(Layer):
-    defaultBases = (zope.ZSERVER_FIXTURE, )
+    defaultBases = (zope.ZSERVER_FIXTURE,)
 
     def testSetUp(self):
         # Start up Selenium
-        driver = os.environ.get('SELENIUM_DRIVER', '').lower() or 'firefox'
+        driver = os.environ.get("SELENIUM_DRIVER", "").lower() or "firefox"
         webdriver = __import__(
-            'selenium.webdriver.{0}.webdriver'.format(driver),
-            fromlist=['WebDriver']
+            f"selenium.webdriver.{driver}.webdriver", fromlist=["WebDriver"]
         )
-        args = [arg.strip() for arg in
-                os.environ.get('SELENIUM_ARGS', '').split()
-                if arg.strip()]
-        self['selenium'] = webdriver.WebDriver(*args)
+        args = [
+            arg.strip()
+            for arg in os.environ.get("SELENIUM_ARGS", "").split()
+            if arg.strip()
+        ]
+        self["selenium"] = webdriver.WebDriver(*args)
 
     def testTearDown(self):
-        self['selenium'].quit()
-        del self['selenium']
+        self["selenium"].quit()
+        del self["selenium"]
 
 
 SELENIUM_FIXTURE = SeleniumLayer()
 SELENIUM_FUNCTIONAL_TESTING = FunctionalTesting(
-    bases=(SELENIUM_FIXTURE, ),
-    name='SeleniumTesting:Functional')
+    bases=(SELENIUM_FIXTURE,), name="SeleniumTesting:Functional"
+)
 SELENIUM_PLONE_FUNCTIONAL_TESTING = FunctionalTesting(
-    bases=(SELENIUM_FIXTURE, PLONE_FIXTURE),
-    name='SeleniumTesting:Functional')
+    bases=(SELENIUM_FIXTURE, PLONE_FIXTURE), name="SeleniumTesting:Functional"
+)
 
 
 # Helper functions
@@ -49,26 +49,25 @@ def open(selenium, url):
 
 
 def login(selenium, portal, username=False, password=False):
-
     if not username:
         username = TEST_USER_NAME
     if not password:
         password = TEST_USER_PASSWORD
 
-    open(selenium, portal.absolute_url() + '/login_form')
-    selenium.find_element_by_name('__ac_name').send_keys(username)
-    selenium.find_element_by_name('__ac_password').send_keys(password)
-    selenium.find_element_by_name('submit').click()
+    open(selenium, portal.absolute_url() + "/login_form")
+    selenium.find_element_by_name("__ac_name").send_keys(username)
+    selenium.find_element_by_name("__ac_password").send_keys(password)
+    selenium.find_element_by_name("submit").click()
 
 
 def click(selenium, xpath):
-    if xpath.count('link='):
-        link = xpath.split('link=')[-1]
+    if xpath.count("link="):
+        link = xpath.split("link=")[-1]
         element = selenium.find_element_by_partial_link_text(link)
-    elif xpath.count('//'):
+    elif xpath.count("//"):
         element = selenium.find_element_by_xpath(xpath)
-    elif xpath.count('#'):
-        eleName = xpath.split('#')[-1]
+    elif xpath.count("#"):
+        eleName = xpath.split("#")[-1]
         element = selenium.find_element_by_id(eleName)
     else:
         element = selenium.find_element_by_name(xpath)
@@ -93,8 +92,8 @@ def typeMce(selenium, value):
     this breaks it won't be a surprise.
     """
     handle = selenium.current_window_handle
-    selenium.switch_to_frame('form.text_ifr')
-    ele = selenium.find_element_by_xpath('//p')
+    selenium.switch_to_frame("form.text_ifr")
+    ele = selenium.find_element_by_xpath("//p")
     ele.send_keys(value)
     selenium.switch_to_window(handle)
 
@@ -103,10 +102,10 @@ def clear(selenium, name):
     selenium.find_element_by_name(name).clear()
 
 
-def select(selenium, xpath1, xpath2=''):
+def select(selenium, xpath1, xpath2=""):
     xpath = xpath1
     if xpath2:
-        xpath = "{0}['{1}']".format(xpath1, xpath2)
+        xpath = f"{xpath1}['{xpath2}']"
         xpath = xpath.replace("select['label=", "select/option['text()=")
     selenium.find_element_by_xpath(xpath).click()
 
@@ -118,7 +117,7 @@ def waitForPageToLoad(selenium, foo):
 
 def publish(selenium):
     click(selenium, "//dl[@id='plone-contentmenu-workflow']/dt/a")
-    click(selenium, '#workflow-transition-publish')
+    click(selenium, "#workflow-transition-publish")
 
 
 def submit(selenium, formId):