+++ src/plone/app/testing/utils.py
@@ -1,12 +1,10 @@
-# -*- coding: utf-8 -*-
 from persistent.list import PersistentList
 from Products.MailHost.MailHost import _mungeHeaders
 from Products.MailHost.MailHost import MailBase
 
 
 class MockMailHost(MailBase):
-    """A MailHost that collects messages instead of sending them.
-    """
+    """A MailHost that collects messages instead of sending them."""
 
     def __init__(self, id):
         self.reset()
@@ -15,13 +13,21 @@ class MockMailHost(MailBase):
         self.messages = PersistentList()
 
     def _send(self, mfrom, mto, messageText, immediate=False):
-        """ Send the message """
+        """Send the message"""
         self.messages.append(messageText)
 
-    def send(self, messageText, mto=None, mfrom=None, subject=None,
-             encode=None, immediate=False, charset=None, msg_type=None):
-        messageText, mto, mfrom = _mungeHeaders(messageText,
-                                                mto, mfrom, subject,
-                                                charset=charset,
-                                                msg_type=msg_type)
+    def send(
+        self,
+        messageText,
+        mto=None,
+        mfrom=None,
+        subject=None,
+        encode=None,
+        immediate=False,
+        charset=None,
+        msg_type=None,
+    ):
+        messageText, mto, mfrom = _mungeHeaders(
+            messageText, mto, mfrom, subject, charset=charset, msg_type=msg_type
+        )
         self.messages.append(messageText)